feat: add arg to only build if input files changed (#60)
If the argument `--ifchanged` is provided, then the outputfile will not be built unless only the layout .yaml file, any of the source files or any of the files in templatepath are newer than the outputfile.
Furthermore, the exit status from the `make` command will be 0 if the outputfile is build, 1 if it is not build and 2 if there was an error.